requirement.txt
Django==1.8
django-annoying==0.10.1
Pillow==3.4.1
pytz==2016.7
